move new hooks to match new file structure
