+++ polymorphia-frontend/hooks/course/projects/useProjectCategories/index.tsx
@@ -2,7 +2,7 @@ import { useQuery } from "@tanstack/react-query";
 import { useEventParams } from "@/hooks/app/params/useEventParams";
 import { EventTypes } from "@/interfaces/general";
 import { ProjectService } from "@/services/project";
-import { UseProjectCategories } from "@/hooks/course/useProjectCategories/types";
+import { UseProjectCategories } from "@/hooks/course/projects/useProjectCategories/types";
 
 export function useProjectCategories(): UseProjectCategories {
   const { gradableEventId, eventType } = useEventParams();
