+++ polymorphia-frontend/hooks/course/projects/useProjectGroupConfiguration/types.ts
