+++ polymorphia-frontend/hooks/course/projects/useProjectGroupUpdate/index.tsx
@@ -2,22 +2,36 @@ import { useMutation, useQueryClient } from "@tanstack/react-query";
 import toast from "react-hot-toast";
 import { useEventParams } from "@/hooks/app/params/useEventParams";
 import {
-  UseProjectGroupDelete,
-  UseProjectGroupDeleteParams,
-} from "@/hooks/course/useProjectGroupDelete/types";
+  UseProjectGroupConfigurationUpdate,
+  UseProjectGroupConfigurationUpdateParams,
+} from "@/hooks/course/projects/useProjectGroupUpdate/types";
 import { ProjectService } from "@/services/project";
 
-export default function useProjectGroupDelete(): UseProjectGroupDelete {
+export default function useProjectGroupUpdate(): UseProjectGroupConfigurationUpdate {
   const { gradableEventId } = useEventParams();
   const queryClient = useQueryClient();
 
-  const mutation = useMutation<void, Error, UseProjectGroupDeleteParams>({
-    mutationFn: async ({ target }) => {
+  const mutation = useMutation<
+    void,
+    Error,
+    UseProjectGroupConfigurationUpdateParams
+  >({
+    mutationFn: async ({ target, configuration }) => {
       return toast.promise(
-        ProjectService.deleteProjectGroup(target, gradableEventId),
+        ProjectService.submitProjectGroupConfiguration(
+          target,
+          gradableEventId,
+          configuration
+        ),
         {
-          loading: "Trwa usuwanie grupy projektowej...",
-          success: "Pomyślnie usunięto grupę projektową.",
+          loading:
+            "Trwa " +
+            (target !== null ? "aktualizowanie" : "tworzenie") +
+            " grupy projektowej...",
+          success:
+            "Pomyślnie " +
+            (target !== null ? "zaktualizowano" : "utworzono") +
+            " grupę projektową.",
         }
       );
     },
